Cart system is linked to users.
The cart class is now used when a user tries to add an item to a cart. Right now no items are added since we haven't figured this part out. For now the addToCart view in views.py will create a cart if the user doesn't have one and will find the users cart to add items to it.